refactor to put create test cases in utility folder
