+++ src/main/python/Linked_List/Remove Nth Node From End of List.py
@@ -12,13 +12,8 @@ The number of nodes in the list is sz.
 0 <= Node.val <= 100
 1 <= n <= sz
 """
+from src.main.python.utility.linked_list import ListNode, initialLinkedList
 
-
-# Definition for singly-linked list.
-# class ListNode:
-#     def __init__(self, val=0, next=None):
-#         self.val = val
-#         self.next = next
 
 class Solution:
     def removeNthFromEnd(self, head: ListNode, n: int) -> ListNode:
