+++ src/main/python/Linked_List/Reverse Linked List.py
@@ -9,13 +9,9 @@ Output: 5->4->3->2->1->NULL
 Follow up:
 A linked list can be reversed either iteratively or recursively. Could you implement both?
 """
+from src.main.python.utility.linked_list import ListNode, initialLinkedList
 
 
-# Definition for singly-linked list.
-class ListNode:
-    def __init__(self, val=0, next=None):
-        self.val = val
-        self.next = next
 class Solution:
     def reverseList_interative(self, head: ListNode) -> ListNode:
         """
